Add integration test
This commit introduces an integration tests verifying
the behavior of the business logic.

In order for the test to call the business logic, most
of the business logic is moved from main.rs to lib.rs

main.rs is just a small wrapper to invoke the business logic.

The integration tests spawns an instance of `Immich`.
That is just a very simple HTTP API server that mimics
the behavior of the 3 HTTP endpoints that the business logic uses.

The test provides a test coverage over 80%.
